Separate avatar remix and get to its own cog and use Nano Banana
